requirements.txt
feedparser==6.0.8
pandas==1.3.5
Pillow==9.0.1
PyGithub==1.55
StackAPI==0.2.0
streamlit==1.5.1
